+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.presentation;
+package kitchenpos.product.ui;
 
 import kitchenpos.product.application.ProductService;
 import kitchenpos.product.domain.Product;
